+++ addons/storyshots/storyshots-puppeteer/package.json
@@ -21,7 +21,7 @@
     "prepare": "node ../../../scripts/prepare.js"
   },
   "dependencies": {
-    "@storybook/core": "5.0.0-beta.3",
+    "@storybook/router": "5.0.0-beta.3",
     "@storybook/node-logger": "5.0.0-beta.3",
     "core-js": "^2.6.2",
     "jest-image-snapshot": "^2.6.0",
