+++ addons/storysource/package.json
@@ -24,6 +24,7 @@
     "@storybook/addons": "5.0.0-beta.3",
     "@storybook/components": "5.0.0-beta.3",
     "@storybook/theming": "5.0.0-beta.3",
+    "@storybook/router": "5.0.0-beta.3",
     "core-js": "^2.6.2",
     "estraverse": "^4.2.0",
     "loader-utils": "^1.2.1",
