+++ lib/client-api/package.json
@@ -22,6 +22,7 @@
     "@storybook/addons": "5.0.0-beta.3",
     "@storybook/client-logger": "5.0.0-beta.3",
     "@storybook/core-events": "5.0.0-beta.3",
+    "@storybook/router": "5.0.0-beta.3",
     "common-tags": "^1.8.0",
     "eventemitter3": "^3.1.0",
     "global": "^4.3.2",
