+++ lib/client-api/src/client_api.js
@@ -4,12 +4,12 @@ import isPlainObject from 'is-plain-object';
 import { logger } from '@storybook/client-logger';
 import addons from '@storybook/addons';
 import Events from '@storybook/core-events';
+import { toId } from '@storybook/router/utils';
 import memoize from 'memoizerific';
 import mergeWith from 'lodash.mergewith';
 import isEqual from 'lodash.isequal';
 
 import subscriptionsStore from './subscriptions_store';
-import toId from './id';
 
 // merge with concatenating arrays, but no duplicates
 const merge = (a, b) =>
@@ -57,7 +57,9 @@ const metaSubscription = () => {
 };
 
 const withSubscriptionTracking = storyFn => {
-  if (!addons.hasChannel()) return storyFn();
+  if (!addons.hasChannel()) {
+    return storyFn();
+  }
   subscriptionsStore.markAllAsUnused();
   subscriptionsStore.register(metaSubscription);
   const result = storyFn();
