+++ lib/client-api/src/index.js
@@ -2,7 +2,6 @@ import ClientApi, { defaultDecorateStory } from './client_api';
 import StoryStore, { splitPath } from './story_store';
 import ConfigApi from './config_api';
 import subscriptionsStore from './subscriptions_store';
-import toId, { sanitize } from './id';
 import pathToId from './pathToId';
 
 export {
@@ -11,8 +10,6 @@ export {
   ConfigApi,
   subscriptionsStore,
   defaultDecorateStory,
-  toId,
-  sanitize,
   pathToId,
   splitPath,
 };
