+++ lib/client-api/src/story_store.js
@@ -4,9 +4,9 @@ import EventEmitter from 'eventemitter3';
 import qs from 'qs';
 import Events from '@storybook/core-events';
 import { logger } from '@storybook/client-logger';
+import { toId } from '@storybook/router/utils';
 import debounce from 'lodash.debounce';
 import { stripIndents } from 'common-tags';
-import toId from './id';
 import pathToId from './pathToId';
 
 // TODO: these are copies from components/nav/lib
