+++ lib/client-api/src/story_store.test.js
@@ -1,9 +1,9 @@
 import { history, document } from 'global';
 import createChannel from '@storybook/channel-postmessage';
 import Events from '@storybook/core-events';
+import { toId } from '@storybook/router/utils';
 
 import StoryStore, { splitPath } from './story_store';
-import toId from './id';
 
 jest.mock('global', () => ({
   history: { replaceState: jest.fn() },
