+++ lib/core/package.json
@@ -31,6 +31,7 @@
     "@storybook/core-events": "5.0.0-beta.3",
     "@storybook/node-logger": "5.0.0-beta.3",
     "@storybook/theming": "5.0.0-beta.3",
+    "@storybook/router": "5.0.0-beta.3",
     "@storybook/ui": "5.0.0-beta.3",
     "airbnb-js-shims": "^1 || ^2",
     "autoprefixer": "^9.4.7",
