+++ lib/core/src/client/preview/index.js
@@ -1,4 +1,4 @@
-import { ClientApi, StoryStore, ConfigApi, toId } from '@storybook/client-api';
+import { ClientApi, StoryStore, ConfigApi } from '@storybook/client-api';
 import start from './start';
 
 export default {
@@ -6,5 +6,4 @@ export default {
   ClientApi,
   ConfigApi,
   StoryStore,
-  toId,
 };
