+++ lib/core/src/client/preview/start.js
@@ -1,7 +1,8 @@
 import addons from '@storybook/addons';
 import { navigator, window, document } from 'global';
 import createChannel from '@storybook/channel-postmessage';
-import { ClientApi, StoryStore, ConfigApi, toId } from '@storybook/client-api';
+import { ClientApi, StoryStore, ConfigApi } from '@storybook/client-api';
+import { toId } from '@storybook/router/utils';
 import { logger } from '@storybook/client-logger';
 import Events from '@storybook/core-events';
 import deprecate from 'util-deprecate';
