+++ lib/router/src/router.tsx
@@ -1,3 +1,4 @@
+import { document } from 'global';
 import React from 'react';
 
 import { Link, Location, navigate, LocationProvider } from '@reach/router';
@@ -55,15 +56,15 @@ interface QueryLinkProps {
   children: React.ReactNode;
 }
 
-const base = document.location.pathname + '?';
+const getBase = () => document.location.pathname + '?';
 
 const queryNavigate = (to: string) => {
-  navigate(`${base}path=${to}`);
+  navigate(`${getBase()}path=${to}`);
 };
 
 // A component that will navigate to a new location/path when clicked
 const QueryLink = ({ to, children, ...rest }: QueryLinkProps) => (
-  <Link to={`${base}path=${to}`} {...rest}>
+  <Link to={`${getBase()}path=${to}`} {...rest}>
     {children}
   </Link>
 );
