+++ lib/router/src/tests/id.test.js
@@ -1,4 +1,4 @@
-import toId from './id';
+import { toId } from '../utils';
 
 describe('toId', () => {
   [
@@ -26,13 +26,11 @@ describe('toId', () => {
 
   it('does not allow story with *no* url chars', () => {
     expect(() => toId('kind', '?')).toThrow(
-      `Invalid story '?', must include alphanumeric characters`
+      `Invalid name '?', must include alphanumeric characters`
     );
   });
 
   it('does not allow empty story', () => {
-    expect(() => toId('kind', '')).toThrow(
-      `Invalid story '', must include alphanumeric characters`
-    );
+    expect(() => toId('kind', '')).toThrow(`Invalid name '', must include alphanumeric characters`);
   });
 });
