+++ lib/ui/src/core/stories.js
@@ -1,6 +1,6 @@
 import mergeWith from 'lodash.mergewith';
 import isEqual from 'lodash.isequal';
-import toId, { sanitize } from '../libs/id';
+import { toId, sanitize } from '@storybook/router/utils';
 
 const merge = (a, b) =>
   mergeWith({}, a, b, (objValue, srcValue) => {
