+++ lib/ui/src/core/url.js
@@ -1,5 +1,5 @@
 import { queryFromLocation } from '@storybook/router';
-import toId from '../libs/id';
+import { toId } from '@storybook/router/utils';
 
 // Initialize the state based on the URL.
 // NOTE:
